adds sweetalert to delete route

+++ src/components/App/App.jsx
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import React from 'react';
+import swal from 'sweetalert';
 import { useEffect, useState } from 'react';
 import Header from '../Header/Header.jsx';
 import GalleryList from '../GalleryList/GalleryList';
@@ -39,14 +40,31 @@ function App() {
 
   const deleteImage = (image) => {
     console.log('Delete from App.jsx')
-    axios.delete(`/gallery/${image.id}`)
+
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this image!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        swal("Poof! Gone forever!", {
+          icon: "success",
+        });
+
+        axios.delete(`/gallery/${image.id}`)
         .then(response => {
             getGallery();
             console.log('successful delete: ', response);
         }).catch(err => {
             console.log('error in delete from App.jsx: ', err);
-        })
-
+        });
+      } else {
+        swal("We'll hang on to this for now!");
+      }
+    });
   }
